+++ noodles-vcf/src/header/record/value/map/contig/builder.rs
@@ -4,6 +4,7 @@ use crate::header::record::value::map::{self, builder::BuildError};
 #[derive(Default)]
 pub struct Builder {
     length: Option<usize>,
+    md5: Option<String>,
     idx: Option<usize>,
 }
 
@@ -11,6 +12,7 @@ impl map::builder::Inner<Contig> for Builder {
     fn build(self) -> Result<Contig, BuildError> {
         Ok(Contig {
             length: self.length,
+            md5: self.md5,
             idx: self.idx,
         })
     }
@@ -27,6 +29,12 @@ impl map::Builder<Contig> {
     /// Sets the length.
     pub fn set_length(mut self, length: usize) -> Self {
         self.inner.length = Some(length);
+        self
+    }
+
+    /// Sets the MD5 hexdigest.
+    pub fn set_md5(mut self, md5: String) -> Self {
+        self.inner.md5 = Some(md5);
         self
     }
 }