+++ noodles-vcf/src/header/record/value/map/contig/tag.rs
@@ -3,11 +3,13 @@ use std::str::FromStr;
 use crate::header::record::value::map::tag::{self, ID, IDX};
 
 const LENGTH: &str = "length";
+const MD5: &str = "md5";
 
 #[derive(Clone, Copy, Debug, Eq, PartialEq)]
 pub enum Standard {
     Id,
     Length,
+    Md5,
     Idx,
 }
 
@@ -18,6 +20,7 @@ impl AsRef<str> for Standard {
         match self {
             Self::Id => ID,
             Self::Length => LENGTH,
+            Self::Md5 => MD5,
             Self::Idx => IDX,
         }
     }
@@ -30,6 +33,7 @@ impl FromStr for Standard {
         match s {
             ID => Ok(Self::Id),
             LENGTH => Ok(Self::Length),
+            MD5 => Ok(Self::Md5),
             IDX => Ok(Self::Idx),
             _ => Err(()),
         }
@@ -44,6 +48,7 @@ mod tests {
     fn test_as_ref_str_for_standard() {
         assert_eq!(Standard::Id.as_ref(), "ID");
         assert_eq!(Standard::Length.as_ref(), "length");
+        assert_eq!(Standard::Md5.as_ref(), "md5");
         assert_eq!(Standard::Idx.as_ref(), "IDX");
     }
 }